+++ lib/src/oauth_web_screen.dart
@@ -88,8 +88,9 @@ class OAuthWebScreen extends StatelessWidget {
     configuration.onCancel?.call();
   }
 
-  Future<bool> onBackPressed(dynamic result) async {
-    if (!((await globalKey.currentState?.onBackPressed(result)) ?? false)) {
+  Future<bool> onBackPressed({dynamic result}) async {
+    if (!((await globalKey.currentState?.onBackPressed(result: result)) ??
+        false)) {
       return false;
     }
     configuration.onCancel?.call();