+++ example/lib/src/auth_sample_screen.dart
@@ -78,7 +78,7 @@ class _AuthSampleScreenState extends State<AuthSampleScreen> {
                             value: e,
                             child: Text(e?.toLanguageTag() ?? 'Default OS')))
                         .toList(),
-                    value: contentLocale,
+                    initialValue: contentLocale,
                     onChanged: (locale) {
                       setState(() {
                         contentLocale = locale;