+++ lib/src/utils/cross_platform_support_noop.dart
@@ -1,4 +1,3 @@
-void loadPage(String url) {
-  // Do nothing
-}
+void loadPage(String url) {/*Do nothing*/}
 String? originUrl() => null;
+void usePathUrlStrategy() {/*Do nothing*/}