+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Types of changes
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## 5.2.0
+### Added
+- Added support for `usePathUrlStrategy()` to clear the path after init. 
+
 ## 5.1.0
 ### Added
 - `userAgent` can now be set in `OAuthConfiguration` and `BaseConfiguration`.